making a parent class of activity holding adaptive banner
